commit pre vulkan rewrite

+++ engine/lib/UUID.cpp
@@ -13,6 +13,6 @@ namespace Slate {
 
     UUID::UUID() : m_UUID(s_UniformDistribution(s_Engine)) {}
 
-    UUID::UUID(uint64_t uuid) : m_UUID(uuid) {}
+    UUID::UUID(uint32_t uuid) : m_UUID(uuid) {}
 
 }
